400, 500,

+++ controllers/index.js
@@ -17,6 +17,92 @@ async function getClient() {
         return
     }
 }
+
+function validatePost(keysArray, data) {
+    if (typeof data != 'object') {
+        return false
+    }
+
+    let dataKeys = []
+    for (i in data) {
+        dataKeys.push(i)
+    }
+
+    let doKeysMatch = keysArray.length == dataKeys.length
+    let keysArrayCopy = [...keysArray]
+    dataKeys.forEach((item) => {
+        if (keysArrayCopy.includes(item)) {
+            keysArrayCopy.splice(keysArrayCopy.indexOf(item), 1)
+        } else {
+            doKeysMatch = false
+        }
+    })
+
+    if (!doKeysMatch) {
+        return false
+    }
+
+    let dataValues = []
+    dataKeys.forEach((item) => {
+        dataValues.push(data[item])
+    })
+
+    let areValuesStrings = dataValues.length == keysArray.length
+    dataValues.forEach((item) => {
+        if (typeof item != 'string') {
+            areValuesStrings = false
+        }
+    })
+
+    if (!areValuesStrings) {
+        return false
+    }
+
+    return true
+}
+
+function validatePut(keysArray, data) {
+    if (typeof data != 'object') {
+        return false
+    }
+    
+    let dataKeys = []
+    for (i in data) {
+        dataKeys.push(i)
+    }
+
+    let doKeysMatch = dataKeys.length >= 1
+    let keysArrayCopy = [...keysArray]
+    dataKeys.forEach((item) => {
+        if (keysArrayCopy.includes(item)) {
+            keysArrayCopy.splice(keysArrayCopy.indexOf(item), 1)
+        } else {
+            doKeysMatch = false
+        }
+    })
+
+    if (!doKeysMatch) {
+        return false
+    }
+
+    let dataValues = []
+    dataKeys.forEach((item) => {
+        dataValues.push(data[item])
+    })
+
+    let areValuesStrings = dataValues.length == keysArray.length
+    dataValues.forEach((item) => {
+        if (typeof item != 'string') {
+            areValuesStrings = false
+        }
+    })
+
+    if (!areValuesStrings) {
+        return false
+    }
+
+    return true
+}
     
 const getAllRoute = async (req, res) => {
     const client = await getClient()
@@ -29,7 +115,12 @@ const getAllRoute = async (req, res) => {
 
     const dataAll = dataParks.concat(dataRestaurants)
 
-    res.send(dataAll)
+    if (dataAll.length >= 1) {
+        res.status(200).send(dataAll)
+    } else {
+        res.status(500).send("Nothing was found.")
+    }
+
     await client.close()
 }
 
@@ -37,28 +128,45 @@ const getAllParksRoute = async (req, res) => {
     const client = await getClient()
     const c = client.db("rexburg").collection("parks").find({})
     const data = await c.toArray()
-    
-    res.send(data)
+
+    if (data.length >= 1) {
+        res.status(200).send(data)
+    } else {   
+        res.status(500).send("Nothing was found.")
+    }
+
     await client.close()
 }
 
 const getAllRestaurantsRoute = async (req, res) => {
     const client = await getClient()
-    const c = client.db("rexburg").collection("restaurants").find({})
-    const data = await c.toArray()
+    const c = client.db("rexburg").collection("restaurants").find({})    
     
-    res.send(data)
+    const data = await c.toArray()
+
+    if (data.length >= 1) {
+        res.status(200).send(data)
+    } else {   
+        res.status(500).send("Nothing was found.")
+    }
+
     await client.close()
 }
 
 const getParkRoute = async (req, res) => {
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
+    
     const data = await client.db("rexburg").collection("parks").findOne({_id: new m.ObjectId(req.params.id)})
     
     if (data) {
-        res.send(data)
+        res.status(200).send(data)
     } else (
-        res.send("Park Not Found")
+        res.status(400).send("Park Not Found")
     )
 
     await client.close()
@@ -66,13 +174,18 @@ const getParkRoute = async (req, res) => {
 }
 
 const getRestaurantRoute = async (req, res) => {
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
     const data = await client.db("rexburg").collection("restaurants").findOne({_id: new m.ObjectId(req.params.id)})
     
     if (data) {
-        res.send(data)
+        res.status(200).send(data)
     } else (
-        res.send("Restaurant Not Found")
+        res.status(400).send("Restaurant Not Found")
     )
 
     await client.close()
@@ -80,49 +193,121 @@ const getRestaurantRoute = async (req, res) => {
 }
 
 const postParkRoute = async (req, res) => {
+    if (!validatePost(["name", "where"], req.body)) {
+        res.status(400).send("The input is invalid.")
+        return
+    }
+
     const client = await getClient()
     data = await client.db("rexburg").collection("parks").insertOne(req.body)
     
-    res.send(data.insertedId)
+    if (data.acknowledged) {
+        res.status(201).send(data.insertedId)
+    } else {
+        res.status(500).send("There was an error in performing this post function.")
+    }
+    
     await client.close()
-
 }
 
 const postResaurantRoute = async (req, res) => {
+    if (!validatePost(["name", "where", "site", "rating", "type", "phone", "orderLink"], req.body)) {
+        res.status(400).send("The input is invalid.")
+        return
+    }
+
     const client = await getClient()
     data = await client.db("rexburg").collection("restaurants").insertOne(req.body)
     
-    res.send(data.insertedId)
+    if (data.acknowledged) {
+        res.status(201).send(data.insertedId)
+    } else {
+        res.status(500).send("There was an error in performing this post function.")
+    }
+    
     await client.close()
-
 }
 
 const putParkRoute = async (req, res) => {
+    if (!validatePut(["name", "where"], req.body)) {
+        res.status(400).send("The body is invalid.")
+        return
+    }
+
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
     put = await client.db("rexburg").collection("parks").updateOne({_id: new m.ObjectId(req.params.id)}, {$set: req.body})
 
-    res.send(`modifiedCount: ${put.modifiedCount}`)
+    if (put.modifiedCount >= 1) {
+        res.status(204).send(`modifiedCount: ${put.modifiedCount}`)
+    } else {
+        res.status(500).send(`There was an error in performing this put function (as modifiedCount is ${put.modifiedCount}).`)
+    }
+
+    await client.close()
 }
 
 const putRestaurantRoute = async (req, res) => {
+    if (!validatePut(["name", "where", "site", "rating", "type", "phone", "orderLink"], req.body)) {
+        res.status(400).send("The body is invalid.")
+        return
+    }
+
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
     put = await client.db("rexburg").collection("restaurants").updateOne({_id: new m.ObjectId(req.params.id)}, {$set: req.body})
 
-    res.send(`modifiedCount: ${put.modifiedCount}`)
+    if (put.modifiedCount >= 1) {
+        res.status(204).send(`modifiedCount: ${put.modifiedCount}`)
+    } else {
+        res.status(500).send(`There was an error in performing this put function (as modifiedCount is ${put.modifiedCount}).`)
+    }
+
+    await client.close()
 }
 
 const deleteParkRoute = async (req, res) => {
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
     del = await client.db("rexburg").collection("parks").deleteOne({_id: new m.ObjectId(req.params.id)})
 
-    res.send(`deletedCount: ${del.deletedCount}`)
+    if (del.deletedCount >= 1) {
+        res.status(200).send(`deletedCount: ${del.deletedCount}`)
+    } else {
+        res.status(500).send(`There was an error in performing this delete function (as deletedCount is ${del.deletedCount}).`)
+    }
+
+    await client.close()
 }
 
 const deleteRestaurantRoute = async (req, res) => {
+    if (!m.ObjectId.isValid(req.params.id)) {
+        res.status(400).send("The given id cannot be used.")
+        return
+    }
+
     const client = await getClient()
     del = await client.db("rexburg").collection("restaurants").deleteOne({_id: new m.ObjectId(req.params.id)})
 
-    res.send(`deletedCount: ${del.deletedCount}`)
+    if (del.deletedCount >= 1) {
+        res.status(200).send(`deletedCount: ${del.deletedCount}`)
+    } else {
+        res.status(500).send(`There was an error in performing this delete function (as deletedCount is ${del.deletedCount}).`)
+    }
+
+    await client.close()
 }
